Move API on route /participants/suggest
