+++ src/backend/defs/rest-api/paths/identities.yaml
@@ -1,43 +1,4 @@
 ---
-identities_suggest:
-  get:
-    description: Returns a list of IDs suggestions according to given parameters/filter. Search is performed within current user's indexes (messages & contacts).
-    tags:
-    - identities
-    security:
-    - basicAuth: []
-    parameters:
-    - name: context
-      in: query
-      description: current user's context (to optimize suggest relevance)
-      required: true
-      type: string
-      enum:
-      - msg_compose
-    - name: q
-      in: query
-      description: a string (3 chars at least) from which to perform the suggestion search
-      required: true
-      type: string
-      minLength: 3
-    produces:
-    - application/json
-    responses:
-      '200':
-        description: a list of suggestions for user to pick-up an identity
-        schema:
-          type: array
-          items:
-            "$ref": "../objects/IdentitySuggest.yaml"
-      '400':
-        description: malform request
-        schema:
-          type: object
-          "$ref": "../objects/Error.yaml"
-      '401':
-        description: Unauthorized access
-        schema:
-          "$ref": "../objects/Error.yaml"
 identities_locals:
   get:
     description: returns the list of user's local identities
@@ -67,7 +28,7 @@ identities_locals:
         schema:
           "$ref": "../objects/Error.yaml"
 
-remotes:
+identities_remotes:
   post:
     description: create a new remote identity for an user
     tags:
