+++ src/backend/defs/rest-api/paths/participants.yaml
@@ -1,0 +1,42 @@
+---
+participants_suggest:
+  get:
+    description: Returns a list of suggestions according to given parameters/filter. Search is performed within current user's indexes (messages & contacts).
+    tags:
+    - participants
+    - suggest
+    security:
+    - basicAuth: []
+    parameters:
+    - name: context
+      in: query
+      description: current user's context (to optimize suggest relevance)
+      required: true
+      type: string
+      enum:
+      - msg_compose
+    - name: q
+      in: query
+      description: a string (3 chars at least) from which to perform the suggestion search
+      required: true
+      type: string
+      minLength: 3
+    produces:
+    - application/json
+    responses:
+      '200':
+        description: a list of suggestions for user to pick-up an identity
+        schema:
+          type: array
+          items:
+            "$ref": "../objects/ParticipantSuggest.yaml"
+      '400':
+        description: malform request
+        schema:
+          type: object
+          "$ref": "../objects/Error.yaml"
+      '401':
+        description: Unauthorized access
+        schema:
+          "$ref": "../objects/Error.yaml"
+
